requirements.txt
streamlit==1.28.1
openai>=1.0.0
anthropic==0.7.0
pydantic==2.4.2
python-dotenv==1.0.0
